move more things to kd plugin
